Self-auditing-RAG/pyproject.toml
[project]
name = "self-auditing-rag"
version = "0.1.0"
description = "Self-Auditing RAG system to detect and reduce hallucinations"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0",
    "faiss-cpu>=1.7",
    "sentence-transformers>=2.2",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "numpy>=1.24",
    "pymupdf>=1.24",
]
